+++ src/styles/main/MainPage.module.css
@@ -15,45 +15,3 @@
   height: 75%;
   position: relative;
 }
-
-.mainPageBtnContainer{
-  width: 100%;
-  height: 10%;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  position: absolute;
-  top: 3%;
-}
-.mainPageDeclarationBtn{
-  width: 100%;
-  height: 10%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  position: absolute;
-  top: 15%;
-}
-.mainPageDeclarationBtnContainer{
-  width: 100%;
-  height: 10%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  position: absolute;
-  bottom: 8%;
-}
-.mainPageDeclarationBtn > div {
-  width: 80%;
-  height: 100%;
-  display: flex;
-  justify-content: end;
-  align-items: center;
-}
-.declarationBtn{
-  z-index: 5;
-}
